Run ubuntu-device-flash in a shorter path
There's a bug with kpartx it seems, that if you are mapping a file with a long path, deleting the mappings can silently fail sometimes. The path that SPI gives us is datestamped, and quite long.  This simply creates the image in a shorter path, then moves it to the proper place so that we don't leak loopback devices.