don't sleep if there's another job in the queue
This patch makes process_jobs try fetching another job, and returning only if
there's none. This way, if there's anything left in the queue, the agent will
start to process it immedietaly.

Previously unit tests mocked only the one return value of what can come from
the server. Logic suggests that when the queue is empty, empty content is
returned. So this patch introduces a fake 'terminator' object and changes
the one returned value to a list of side effects (two - first -> fake content,
second -> the terminator).

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>