Detect if we have empty relation data and exit
There's a chance that the relation can exist but not yet have data in
it. This can result in an error if we try to use that data later, and
the charm will get stuck in an error state rather than retrying. Exiting
here if there's no data for the relation forces it to retry getting the
data later.