Muxpi improvement (#227)
* Introduce a new parameter 'media' in 'provision_data' section

This parameter is designed to be used with 'sd' or 'usb', matching 'sd-disk' or 'tc-disk' symlink dev names, to make sure the device name is always predictable (when 2 boot media are visible by a controller, device names are randomly related to one or the other). A specific udev rule is necessary (simlarly to what's installed by git+ssh://git.launchpad.net/~riverside-team/riverside/+git/zapper-setup)

* Replace match statement with if/elif/else to ensure compatibility

* Updated documentation for new parameter

* Use latest 'plug_to_self' zapper subcommand to get 'test_device' node

Using symlinks isn't necessary anymore thanks to this new sub command on zapper. That avoid the pain of installing udev rules as a pre-requisite

* Fix typo

* Reformat comment to keep line shorter than 80