fix: remove attachment data after unpacking to eliminate interference
The provisioning and firmware update phases will be triggered when
`provision_data` or `firmware_update` data are not empty, which is the
case when attachments are provided for these phases.

In order to minimise changes to the agent code and keep its logic
exactly the same as before (for now), attachment data is *stripped*
from the job data dict as soon as the attachments are unpacked. This
eliminates any interference that the presence of attachments may have
to the execution of each phase.