timeout muxpi hdparm after 40s to workaround usb reset
On some zapper-connected devices, there's an occasional usb reset that
seems to happen while running hdparm -z. When this happens, it makes the
hdparm command take about 32s when it normally doesn't even take 1s.
Since we aren't seeing other negative effects from this so far, let's
extend the timeout a bit so that provisioning works while we debug this
further