Log the error and fail gracefully for artifact download issues (#321)
If the disk fills up during the job, it's not surprising that things can
fail quite badly. But then to finish off the job, we try to save the
artifacts, which is also probably going to fail. Let's log as much
detail about the failure as we can, salvage whatever we can to push as a
result to the server, and clean it up so that we fail more gracefully if
this happens.