fix(serial): Write log file from TCP in binary format
By opening the serial log file in binary ("b") mode, we can
write the data received from the TCP socket directly to the
file without having to (try to) decode the content as UTF-8.

This is especially important if a valid UTF-8 character happens
to cross the boundary of a 4096-byte read, resulting in lost data
instead of a single, valid UTF-8 character in the file.