Add agent_name in TF environment (#354)
* Add agent_name in TF environment

It is currently impossible to guess the name of the device agent from a testflinger job running from a device agent queue. This is unfortunate since we have a few Riverside tools that abstract the complexity of a setup, given the device name as an argument. Adding it as an environment variable would reduce the development effort on those tools and can probably benefit as well to other use cases

* Use uppercase env variable (AGENT_NAME), and add a default value

Co-authored-by: Paul Larson <pwlars@gmail.com>

* Fix get configuration command

---------

Co-authored-by: Paul Larson <pwlars@gmail.com>